feat: provide detailed docs & fix: tailwind affecting docusaurus pages
