+++ website/sidebars.js
@@ -1,5 +1,6 @@
 module.exports = {
   main: {
+    "Getting Started": ["why_use", "installation"],
     CLI: ["cli"],
     Congfiguration: ["config"],
   },
