+++ website/src/css/custom.css
@@ -23,3 +23,11 @@
   margin: 0 calc(-1 * var(--ifm-pre-padding));
   padding: 0 var(--ifm-pre-padding);
 }
+
+.homebody {
+  padding: 40px;
+}
+
+.card {
+  background-color: rgb(245, 245, 245);
+}
