+++ website/tailwind.config.js
@@ -1,6 +1,7 @@
 module.exports = {
   purge: ["./src/**/*.html", "./src/**/*.js", "./src/**/*.tsx"],
   darkMode: "media", // or 'media' or 'class'
+  important: "#tailwind",
   theme: {
     extend: {},
   },
